# doctor: use multithread to accelerate speed