Create two type of threads:
 * worker: do the actual check and send check result to report thread.
 * reporter: print the warning message to stderr.

Benchmark for `129 kegs, 71,153 files, 4.5G` Cellar.
 * Old: `brew doctor  3.61s user 0.87s system 22% cpu 20.001 total`
 * New: `brew doctor  4.35s user 0.96s system 38% cpu 13.880 total`

Note: the performance will vary based on the diagnostic result.
Under worst case, there will be no difference between multithread
model and original sequential model.

cc @UniqMartin @mikemcquaid