@xu-cheng 2016-01-06
As mentioned in other PR, there isn't a significant improvement.  I think the reason is GIL. I will see if possible to make it multiprocess instead of multithread.