@mikemcquaid 2016-01-06
> I think the reason is GIL.

Either that or it's just IO-bound.