@UniqMartin 2016-01-07
@xu-cheng Thanks for giving it a try and disproving my claim! :bow:

> Either that or it's just IO-bound.

I tried playing with the number of workers, but without significant improvement. IO-bound sounds quite plausible, given that `check_for_linked_keg_only_brews` is by far the slowest check. Thus parallelism might actually degrade performance.

Benchmark for a `163 kegs, 80,632 files, 2.4G` cellar on a Retina MacBook Pro (13-inch, Late 2012):

```
# serial code, slow checks last
brew doctor:  10.880s user,   0.930s system,  15.142s total ( 78.0% cpu)
brew doctor:  10.890s user,   0.990s system,  16.100s total ( 73.8% cpu)

# parallelized, slow checks last
brew doctor:  11.650s user,   1.010s system,  12.126s total (104.4% cpu)
brew doctor:  13.400s user,   1.130s system,  14.041s total (103.5% cpu)

# parallelized, slow checks first
brew doctor:  11.610s user,   1.340s system,  12.224s total (105.9% cpu)
brew doctor:  13.020s user,   1.110s system,  13.682s total (103.3% cpu)
```

Thoughts:

- Multi-threading should probably be disabled for `-D` and invocation with named test(s).
- Non-deterministic output order might be quite bad from a UX perspective, especially for people with rather messy setups.
- Results so far don't seem good enough to warrant the added complexity and worse UX.