@mikemcquaid 2016-01-07
> Results so far don't seem good enough to warrant the added complexity and worse UX.

Agreed. I'd like to see at least a 50% speedup to make the code complexity worthwhile. There's possibly ways we can speed things up by doing less disk access or e.g. using something like `find` to get some of the results.