@xu-cheng 2016-01-07
Yeah, this PR is most for experiment purpose. Close for now.