@UniqMartin 2016-01-07
> There's possibly ways we can speed things up by doing less disk access

If I find the time, I'll have a look at the biggest offenders and see if they can be improved.

> or e.g. using something like `find` to get some of the results.

I'm personally against doing this, because the price of the speedup (if there's any) is that we would have to rely on external tools that can break in more ways than native Ruby code (due to user configuration, overrides in `PATH`,  etc.) and that we would have to parse the textual output of those tools (which comes with its own set of corner cases). The benefit has to be pretty significant to offset these pains.