@mikemcquaid 2016-01-09
> I'm personally against doing this, because the price of the speedup (if there's any) is that we would have to rely on external tools that can break in more ways than native Ruby code (due to user configuration, overrides in PATH, etc.) and that we would have to parse the textual output of those tools (which comes with its own set of corner cases). The benefit has to be pretty significant to offset these pains.

Obviously I'm only in favour if there's a fairly dramatic speedup. As this is a standard Unix tool the output parsing is extremely straightforward. I've certainly seen "iterate over files in Ruby" massively sped up in the past by letting `find` do the same thing.