@UniqMartin 2016-01-09
> As this is a standard Unix tool the output parsing is extremely straightforward.

The parsing is never straightforward thanks to encoding issues, special characters like newlines in file names (which are perfectly legal on OS X), and other issues I probably have forgotten. Working with the functionality that a programming language offers at least shields you from this (or gives you someone to blame if that fails, other than yourself).